# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing reservations",
        "title": "Reservation API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/discount-codes": {
            "get": {
                "description": "Get all discount codes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discount-codes"
                ],
                "summary": "Gets all discount codes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.GetAllDiscountCodesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/reservations": {
            "post": {
                "description": "Creates a new reservation with selected seat IDs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Create a new reservation",
                "parameters": [
                    {
                        "description": "Reservation Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateReservationDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateReservationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/reservations/cancel": {
            "post": {
                "description": "Cancel given reservation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Cancel existing reservation",
                "parameters": [
                    {
                        "description": "Reservation Id to cancel",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CancelReservationStatusDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BoolResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/reservations/confirm": {
            "post": {
                "description": "Confirm given reservation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Confirm existing reservation",
                "parameters": [
                    {
                        "description": "Reservation Id to confirm",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.ConfirmReservationStatusDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BoolResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/reservations/is-seat-reserved": {
            "get": {
                "description": "Checks if given seat is reserved",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Checks if given seat is reserved",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Seat ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BoolResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/reservations/{id}": {
            "get": {
                "description": "Returns reservation information and associated seats",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Get reservation details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reservation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.ReservationSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/seats": {
            "get": {
                "description": "Get all seats",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "seats"
                ],
                "summary": "Gets all seats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.GetAllSeatsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new seat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "seats"
                ],
                "summary": "Creates a new seat",
                "parameters": [
                    {
                        "description": "Seat Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateSeatDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateSeatResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.ReservationStatus": {
            "type": "string",
            "enum": [
                "pending",
                "confirmed",
                "expired"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusConfirmed",
                "StatusExpired"
            ]
        },
        "dtos.CancelReservationStatusDto": {
            "type": "object",
            "required": [
                "reservationId"
            ],
            "properties": {
                "reservationId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "dtos.ConfirmReservationStatusDto": {
            "type": "object",
            "required": [
                "reservationId"
            ],
            "properties": {
                "reservationId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "dtos.CreateReservationDto": {
            "type": "object",
            "required": [
                "seatIds"
            ],
            "properties": {
                "seatIds": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dtos.CreateReservationResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 201
                },
                "data": {
                    "type": "string",
                    "example": "created-reservation-uuid"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "dtos.CreateSeatDto": {
            "type": "object",
            "required": [
                "col",
                "price",
                "row"
            ],
            "properties": {
                "col": {
                    "type": "integer",
                    "minimum": 1
                },
                "price": {
                    "type": "number",
                    "minimum": 1
                },
                "row": {
                    "type": "string",
                    "maxLength": 1
                }
            }
        },
        "dtos.CreateSeatResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 201
                },
                "data": {
                    "type": "string",
                    "example": "created-seat-uuid"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "dtos.GetAllDiscountCodesResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.DiscountCode"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "dtos.GetAllSeatsResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Seat"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "dtos.ReservationDto": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "price": {
                    "type": "number",
                    "example": 35
                },
                "seats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.SeatDto"
                    }
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.ReservationStatus"
                        }
                    ],
                    "example": "confirmed"
                }
            }
        },
        "dtos.ReservationSuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/dtos.ReservationDto"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "dtos.SeatDto": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "label": {
                    "type": "string",
                    "example": "A1"
                }
            }
        },
        "entities.DiscountCode": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "BIG_FAM"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "discount-code-uuid"
                },
                "price": {
                    "type": "number",
                    "example": 10
                }
            }
        },
        "entities.Seat": {
            "type": "object",
            "properties": {
                "col": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "string",
                    "example": "seat-uuid"
                },
                "price": {
                    "type": "number",
                    "example": 15
                },
                "row": {
                    "type": "string",
                    "example": "A"
                }
            }
        },
        "response.BoolResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "boolean",
                    "example": true
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "error message"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        }
    }
}